// src/app/api/swagger/models/api-market-deal.ts
/* tslint:disable */
/* eslint-disable */
import { ApiApprovalRequestStateEnum } from '../models/api-approval-request-state-enum';
import { ApiDictionaryItemAltDto } from '../models/api-dictionary-item';
import { ApiDiscussionItemTypeEnum } from '../models/api-discussion-item-type-enum';
import { ApiMarketDealStateEnum } from '../models/api-market-deal-state-enum';
import { ApiPurchaseMethodTypeEnum } from '../models/api-purchase-method-type-enum';
import { ApiRetradingTypeEnum } from '../models/api-retrading-type-enum';
import { ApiTradeBusinessFlowEnum } from '../models/api-trade-business-flow-enum';
export interface ApiMarketDealAltDto {
  ApprovalRequestState?: ApiApprovalRequestStateEnum | null;
  AuctionEndDate?: string | null;
  BidsCount?: number | null;
  BusinessFlowType?: ApiTradeBusinessFlowEnum | null;
  CountApplications?: number;
  CounterPartyOrganizationGuid?: string;
  CounterPartyOrganizationId?: number;
  CurrencyCode?: string | null;
  CustomerExternalOrganizationId?: string | null;
  CustomerGuid?: string;
  CustomerNameAndInn?: string | null;
  CustomerSource?: number | null;
  DealStateId?: number;
  DealStateName?: string | null;
  DeliveryKladrRegionName?: string | null;
  DeliveryKladrs?: Array<ApiDictionaryItemAltDto> | null;
  DiscussionId?: number | null;
  Downloads?: number | null;
  ExternalId?: string | null;
  ExternalSystemId?: number | null;
  FillingApplicationEndDate?: string | null;
  Hosts?: Array<string> | null;
  Id?: number;
  IsAnyApplicationExists?: boolean;
  IsContractSigningSuspended?: boolean;
  IsDetailedMetricsAccessible?: boolean;
  IsImmediate?: boolean;
  IsImportPhaseout?: boolean | null;
  IsPublicApplication?: boolean;
  IsRetradingEnabled?: boolean;
  IsSuspended?: boolean;
  IsUnitBidding?: boolean;
  LastDiscussionItemDate?: string | null;
  LastDiscussionItemType?: ApiDiscussionItemTypeEnum | null;
  LotItemsCount?: number | null;
  MarketDealState?: ApiMarketDealStateEnum;
  MarketDealStateDescription?: string | null;
  OnlyOneRetradingBidAllowed?: boolean;
  OrderChangeQuantity?: number | null;
  OrderChangeUnitPrice?: number | null;
  OrganizerGuid?: string;
  OrganizerNameAndInn?: string | null;
  PictureFileGuids?: Array<string> | null;
  PlanNumber?: string | null;
  PlanPositionNumber?: string | null;
  Price?: number | null;
  PublicationDate?: string | null;
  PurchaseMethodType?: ApiPurchaseMethodTypeEnum;
  Quantity?: number | null;
  RetradingEndDate?: string | null;
  RetradingStartDate?: string | null;
  RetradingType?: ApiRetradingTypeEnum | null;
  ShortOkeiName?: string | null;
  TradeName?: string | null;
  TradeNumber?: string | null;
  UnitPrice?: number | null;
  UnitPriceInCurrency?: number | null;
  Views?: number | null;
}